Add CommandPlayer to manager command execution
